tests/wpt/tests/html/canvas/element/manual/filters/svg-filter-feImage-cross-origin-taint.https.sub.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>SVG feImage filter using a cross origin image should taint the canvas</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<svg width="100" height="100">
  <filter id="filter-cross">
    <feImage href="https://{{hosts[alt][www]}}:{{ports[https][0]}}/images/red-100x100.png"/>
  </filter>
</svg>
<canvas id=canvas width=100 height=100></canvas>
<script>
  promise_test(async function() {
    const preloadImage = (url) => new Promise((resolve, reject) => {
      const img = new Image();
      img.onload = () => resolve(img);
      img.onerror = () => reject(new Error(`Failed to load image: ${url}`));
      img.src = url;
    });

    const feImageEl = document.querySelector('#filter-cross feimage');
    await Promise.all([
      preloadImage(feImageEl.getAttribute('href')),
    ]);

    context = canvas.getContext("2d");
    context.fillStyle = "red";
    context.filter = "url(#filter-cross)";
    context.fillRect(0, 0, 100, 100);

    // The canvas should be tainted, so the following should throw.
    assert_throws_dom("SecurityError",
      function () { canvas.toDataURL() },
      'canvas.toDataURL should throw SecurityError for SVG using a cross origin image filter.'
    );
  }, "Tainting check for cross origin image in SVG feImage filter");
</script>
